mass polymorph change b

+++ include/EXTLLVM.h
@@ -82,7 +82,7 @@ unsigned long string_hash(unsigned char* str);
   int llvm_sprintf(char* str, char* format, ...);
   void llvm_send_udp(char* host, int port, void* message, int message_length);
   double imp_randd();
-  double imp_randf();
+  float imp_randf();
   int64_t imp_rand1(double a);
   int64_t imp_rand2(double a, double b);
